分类整理无线接口类以及记录类
