+++ src/main/java/cn/qiuye/gtmoremachine/api/misc/wireless/cwu/WirelessCWUContainer.java
@@ -1,5 +1,7 @@
 package cn.qiuye.gtmoremachine.api.misc.wireless.cwu;
 
+import cn.qiuye.gtmoremachine.api.misc.wireless.cwu.record.BasicTransferData;
+import cn.qiuye.gtmoremachine.api.misc.wireless.cwu.Interface.ITransferData;
 import cn.qiuye.gtmoremachine.api.misc.wireless.time.TimeStat;
 import cn.qiuye.gtmoremachine.data.wireless.cwu.WirelessCWUSavedData;
 import cn.qiuye.gtmoremachine.utils.BigIntegerUtils;
