+++ src/main/java/cn/qiuye/gtmoremachine/api/misc/wireless/energy/Interface/ITransferData.java
@@ -1,8 +1,10 @@
-package cn.qiuye.gtmoremachine.api.misc.wireless.energy;
+package cn.qiuye.gtmoremachine.api.misc.wireless.energy.Interface;
 
+import cn.qiuye.gtmoremachine.api.gui.monitor.Format;
+import cn.qiuye.gtmoremachine.utils.FormattingUtil;
+import cn.qiuye.gtmoremachine.utils.NumberUtils;
 import cn.qiuye.gtmoremachine.utils.TeamUtils;
 
-import com.gregtechceu.gtceu.api.GTValues;
 import com.gregtechceu.gtceu.api.machine.MetaMachine;
 
 import com.lowdragmc.lowdraglib.gui.widget.ComponentPanelWidget;
@@ -11,19 +13,21 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.HoverEvent;
 import net.minecraft.network.chat.Style;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
-public interface IDimensionTransferData {
+public interface ITransferData {
 
     UUID UUID();
 
-    int Voltagelevel();
+    BigInteger Throughput();
 
     MetaMachine machine();
 
-    default Component getInfo() {
+    default Component getInfo(Format format) {
         MetaMachine machine = machine();
-        int Voltagelevel = Voltagelevel();
+        BigDecimal eut = new BigDecimal(Throughput());
         String pos = machine.getBlockPos().toShortString();
         return Component.translatable(machine.getBlockState().getBlock().getDescriptionId())
                 .withStyle(Style.EMPTY.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT,
@@ -31,7 +35,10 @@ public interface IDimensionTransferData {
                                 machine.getLevel().dimension().location()).append(" [").append(pos).append("] ")
                                 .append(Component.translatable("gtmoremachine.machine.wireless_monitor.tooltip.0",
                                         TeamUtils.getName(machine.getLevel(), UUID()))))))
-                .append(GTValues.VNF[Voltagelevel])
+                .append((eut.compareTo(BigDecimal.ZERO) > 0 ? " +" : " ") + NumberUtils.formatBigDecimalNumberOrSic(eut, format))
+                .append(" EU/t (")
+                .append(NumberUtils.formatBigDecimalNumberOrSic(FormattingUtil.voltageAmperage(eut), format) + " A ")
+                .append(FormattingUtil.voltageName(eut)).append(")")
                 .append(ComponentPanelWidget.withButton(Component.literal(" [ ] "), pos));
     }
 }
