+++ src/main/java/cn/qiuye/gtmoremachine/api/misc/wireless/energy/record/CapacityStorageData.java
@@ -1,5 +1,6 @@
-package cn.qiuye.gtmoremachine.api.misc.wireless.energy;
+package cn.qiuye.gtmoremachine.api.misc.wireless.energy.record;
 
+import cn.qiuye.gtmoremachine.api.misc.wireless.energy.Interface.ICapacitylimitData;
 import com.gregtechceu.gtceu.api.machine.MetaMachine;
 
 import java.math.BigInteger;
