+++ src/main/java/cn/qiuye/gtmoremachine/api/misc/wireless/energy/record/StoragePercentageData.java
@@ -1,4 +1,4 @@
-package cn.qiuye.gtmoremachine.api.misc.wireless.energy;
+package cn.qiuye.gtmoremachine.api.misc.wireless.energy.record;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
