+++ src/main/java/cn/qiuye/gtmoremachine/common/item/WirelessEnergyTerminalBehavior.java
@@ -6,7 +6,7 @@ import cn.qiuye.gtmoremachine.api.gui.widget.AlignComponentPanelWidget;
 import cn.qiuye.gtmoremachine.api.gui.widget.AlignLabelWidget;
 import cn.qiuye.gtmoremachine.api.item.ModularHUD;
 import cn.qiuye.gtmoremachine.api.machine.IWirelessEnergyContainerHolder;
-import cn.qiuye.gtmoremachine.api.misc.wireless.energy.IWirelessMonitor;
+import cn.qiuye.gtmoremachine.api.misc.wireless.energy.Interface.IWirelessMonitor;
 import cn.qiuye.gtmoremachine.api.misc.wireless.energy.WirelessEnergyContainer;
 import cn.qiuye.gtmoremachine.utils.FormattingUtil;
 import cn.qiuye.gtmoremachine.utils.NumberUtils;
