+++ src/main/java/cn/qiuye/gtmoremachine/common/machine/electric/WirelessEnergyMonitor.java
@@ -3,7 +3,7 @@ package cn.qiuye.gtmoremachine.common.machine.electric;
 import cn.qiuye.gtmoremachine.api.gui.monitor.*;
 import cn.qiuye.gtmoremachine.api.gui.widget.AlignComponentPanelWidget;
 import cn.qiuye.gtmoremachine.api.gui.widget.AlignLabelWidget;
-import cn.qiuye.gtmoremachine.api.misc.wireless.energy.IWirelessMonitor;
+import cn.qiuye.gtmoremachine.api.misc.wireless.energy.Interface.IWirelessMonitor;
 import cn.qiuye.gtmoremachine.api.misc.wireless.energy.WirelessEnergyContainer;
 
 import com.gregtechceu.gtceu.api.blockentity.BlockEntityCreationInfo;
